refactor: move connection dropdown to own file
This moves ChangeConnectionBlockModal to it's own file so it's easier to
manage.
